Fix path interference issue between pysqream and pysqream_blue
This commit addresses the problem of interference between the pysqream
and pysqream_blue folders when both were added to the PATH.
The interference occurred in cases where absolute imports were used,
leading to conflicts and confusion.

Previously, the pysqream.connection module had the line from ping
import PingLoop, while the pysqream_blue module also had a ping module
with a PingLoop (pretend they are different). Both folders were added
to the PATH, making them visible to Python during module imports.

As a result, when importing PingLoop from pysqream.connection, it
inadvertently imported PingLoop from pysqream_blue.ping due to the
presence of the pysqream_blue folder in the PATH. This example
illustrates a simple case, but the issue extended to other modules and
caused confusion.

This commit removes the inclusion of the whole folder of pysqream in
the PATH, resolving the interference problem between pysqream and
pysqream_blue at pysqream side.

After that all absolute import should be replaced in package by
relative imports